apply fragment structure to menu activity
