--- app/src/main/res/layout/activity_choose_difficulty.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.kalantos.spitball.activities.ChooseDifficultyActivity"
-    android:background="@drawable/backgreen">
-
-    <Button
-        android:layout_width="100dp"
-        android:layout_height="60dp"
-        android:text="@string/forDummies"
-        android:id="@+id/button"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="48dp"
-        android:onClick="forDummiesDifficult"/>
-
-    <Button
-        android:layout_width="100dp"
-        android:layout_height="60dp"
-        android:text="@string/easy"
-        android:id="@+id/button5"
-        android:layout_centerVertical="true"
-        android:layout_centerHorizontal="true"
-        android:onClick="easyDifficult"
-        />
-
-    <Button
-        android:layout_width="100dp"
-        android:layout_height="60dp"
-        android:text="@string/hard"
-        android:id="@+id/button6"
-        android:layout_alignParentBottom="true"
-        android:layout_alignStart="@+id/button5"
-        android:layout_marginBottom="44dp"
-        android:onClick="hardDifficult"/>
-</RelativeLayout>
