--- app/src/main/res/layout/activity_choose_type_of_game.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_choose_type_of_game"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.kalantos.spitball.activities.ChooseTypeOfGameActivity"
-    android:background="@drawable/backgreen">
-
-
-    <Button
-        android:text="Online"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/button9"
-        android:onClick="createOnlineGame"
-        android:layout_alignParentBottom="true"
-        android:layout_alignStart="@+id/button8"
-        android:layout_marginBottom="105dp" />
-
-    <Button
-        android:text="Local"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/button8"
-        android:onClick="intentGame"
-        android:layout_above="@+id/button9"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="38dp" />
-
-    <ImageView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_above="@+id/button9"
-        android:layout_alignParentStart="true"
-        android:layout_marginStart="56dp"
-        android:layout_marginBottom="17dp"
-        android:id="@+id/imageView11" />
-</RelativeLayout>
