+++ app/src/main/res/layout/activity_menu.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -7,80 +7,12 @@
     tools:context="com.kalantos.spitball.activities.MenuActivity"
     android:background="@drawable/backgreen">
 
-
-
-    <Button
-        android:layout_width="100dp"
-        android:layout_height="50dp"
-        android:text="@string/players2"
-        android:id="@+id/button2"
-        android:layout_above="@+id/button3"
-        android:layout_alignStart="@+id/button3"
-        android:onClick="intentChooseTypeOfGame"/>
-
-    <Button
-        android:layout_width="100dp"
-        android:layout_height="50dp"
-        android:text="@string/player1"
-        android:id="@+id/button3"
-        android:layout_centerVertical="true"
-        android:layout_centerHorizontal="true"
-        android:onClick="intentChooseDifficulty"/>
-    <Button
-        android:layout_width="100dp"
-        android:layout_height="50dp"
-        android:text="@string/highScores"
-        android:id="@+id/button4"
-        android:layout_below="@+id/button3"
-        android:layout_alignStart="@+id/button3"
-        android:onClick="intentHighScores"/>
-
-    <ImageView
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:src="@drawable/ballgreen"
-        android:id="@+id/imageView0"
-        android:layout_above="@+id/imageView2"
-        android:layout_alignParentStart="true" />
-    <ImageView
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:src="@drawable/ballgreen"
-        android:id="@+id/imageView1"
-        android:layout_alignTop="@+id/button2"
-        android:layout_toEndOf="@+id/imageView0" />
-    <ImageView
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:src="@drawable/ballgreen"
-        android:id="@+id/imageView2"
-        android:layout_alignTop="@+id/button3"
-        android:layout_alignEnd="@+id/imageView1"
-        android:layout_marginEnd="46dp" />
-    <ImageView
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:src="@drawable/ballpink"
-        android:id="@+id/imageView4"
-        android:layout_marginStart="60dp"
-        android:layout_below="@+id/button2"
-        android:layout_toEndOf="@+id/button2" />
-
-    <ImageView
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:src="@drawable/ballpink"
-        android:id="@+id/imageView3"
-        android:layout_below="@+id/button3"
-        android:layout_toEndOf="@+id/imageView5" />
-
-    <ImageView
-        android:layout_width="100dp"
-        android:layout_height="100dp"
-        android:src="@drawable/ballpink"
-        android:id="@+id/imageView5"
-        android:layout_below="@+id/imageView4"
-        android:layout_toEndOf="@+id/button4" />
+    <LinearLayout
+        android:id="@+id/fragmentHolderMenu"
+        android:orientation="vertical"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+    </LinearLayout>
 
     <ImageView
         android:layout_width="40dp"
@@ -88,16 +20,8 @@
         android:src="@drawable/settings"
         android:id="@+id/imageView6"
         android:onClick="intentSettings"
-        android:layout_above="@+id/imageView0"
-        android:layout_toEndOf="@+id/imageView3" />
+        android:layout_gravity="end"
+        android:layout_marginEnd="20dp"
+        android:layout_marginTop="20dp"/>
 
-    <TextView
-        android:text="TextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_above="@+id/imageView0"
-        android:layout_centerHorizontal="true"
-        android:id="@+id/textView3" />
-
-
-</RelativeLayout>
+</FrameLayout>
