+++ app/src/main/res/layout/fragment_choose_difficulty.xml
@@ -1,13 +1,44 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.kalantos.spitball.fragments.ChooseDifficultyFragment">
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+xmlns:android="http://schemas.android.com/apk/res/android"
+xmlns:app="http://schemas.android.com/apk/res-auto"
+xmlns:tools="http://schemas.android.com/tools"
+android:orientation="vertical"
+android:layout_width="match_parent"
+android:layout_height="match_parent">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+<Button
+    android:text="@string/forDummies"
+    android:layout_width="100dp"
+    android:layout_height="50dp"
+    android:id="@+id/button10"
+    android:layout_alignParentTop="true"
+    android:layout_alignParentStart="true"
+    android:layout_marginTop="112dp"
+    app:layout_constraintTop_toTopOf="parent"
+    android:layout_marginStart="16dp"
+    app:layout_constraintLeft_toLeftOf="parent"
+    android:layout_marginEnd="16dp"
+    app:layout_constraintRight_toRightOf="parent"
+    android:onClick="forDummiesDifficult"/>
 
-</FrameLayout>
+<Button
+    android:text="@string/easy"
+    android:layout_width="100dp"
+    android:layout_height="50dp"
+    android:id="@+id/button11"
+    android:layout_marginTop="32dp"
+    app:layout_constraintTop_toBottomOf="@+id/button10"
+    app:layout_constraintLeft_toLeftOf="@+id/button10"
+    android:onClick="easyDifficult"/>
+
+<Button
+    android:text="@string/hard"
+    android:layout_width="100dp"
+    android:layout_height="50dp"
+    android:id="@+id/button12"
+    android:layout_marginTop="32dp"
+    app:layout_constraintTop_toBottomOf="@+id/button11"
+    app:layout_constraintLeft_toLeftOf="@+id/button11"
+    android:onClick="hardDifficult"/>
+</android.support.constraint.ConstraintLayout>
