+++ app/src/main/res/layout/fragment_choose_type_of_game.xml
@@ -1,13 +1,36 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.kalantos.spitball.fragments.ChooseTypeOfGameFragment">
+    android:layout_height="match_parent">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+    <Button
+        android:text="Online"
+        android:layout_width="100dp"
+        android:layout_height="50dp"
+        android:id="@+id/button10"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentStart="true"
+        android:layout_marginTop="160dp"
+        app:layout_constraintTop_toTopOf="parent"
+        android:layout_marginStart="16dp"
+        app:layout_constraintLeft_toLeftOf="parent"
+        android:layout_marginEnd="16dp"
+        app:layout_constraintRight_toRightOf="parent"
+        android:onClick="intentChooseTypeOfGame"
+        />
 
-</FrameLayout>
+    <Button
+        android:text="Local"
+        android:layout_width="100dp"
+        android:layout_height="50dp"
+        android:id="@+id/button11"
+        android:layout_marginTop="32dp"
+        app:layout_constraintTop_toBottomOf="@+id/button10"
+        app:layout_constraintLeft_toLeftOf="@+id/button10"
+        android:onClick="intentGame"/>
+
+</android.support.constraint.ConstraintLayout>
