--- app/src/main/res/layout/fragment_how_to_play1.xml
@@ -1,24 +1,0 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.kalantos.spitball.activities.howToPlayActivity">
-
-    <!-- TODO: Update blank fragment layout -->
-
-    <ImageView
-        android:layout_width="221dp"
-        android:layout_height="134dp"
-        android:id="@+id/imageView7"
-        android:layout_gravity="left|center_vertical"
-        android:src="@drawable/ballgreen"/>
-
-    <TextView
-        android:layout_width="361dp"
-        android:layout_height="199dp"
-        android:textAppearance="?android:attr/textAppearanceLarge"
-        android:text="@string/howToPlayInfo"
-        android:id="@+id/textView2"
-        android:layout_gravity="right|center_vertical"
-        android:textColor="@color/abc_input_method_navigation_guard" />
-</FrameLayout>
