--- app/src/main/res/layout/fragment_how_to_play3.xml
@@ -1,15 +1,0 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.kalantos.spitball.fragments.HowToPlay3">
-
-    <ImageView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/imageView9"
-        android:layout_gravity="center"
-        android:src="@drawable/howto3"
-        />
-
-</FrameLayout>
