--- app/src/main/res/layout/fragment_how_to_play4.xml
@@ -1,15 +1,0 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.kalantos.spitball.fragments.HowToPlay4">
-
-    <ImageView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/imageView10"
-        android:layout_gravity="center"
-        android:src="@drawable/howto4"
-        />
-
-</FrameLayout>
