+++ app/src/main/res/layout/fragment_menu.xml
@@ -1,38 +1,45 @@
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.kalantos.spitball.fragments.MenuFragment">
-
+    android:layout_height="match_parent">
 
     <Button
-        android:text="Button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/button3"
+        android:text="@string/players2"
+        android:layout_width="100dp"
+        android:layout_height="50dp"
+        android:id="@+id/button10"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentStart="true"
         android:layout_marginTop="112dp"
         app:layout_constraintTop_toTopOf="parent"
         android:layout_marginStart="16dp"
         app:layout_constraintLeft_toLeftOf="parent"
         android:layout_marginEnd="16dp"
-        app:layout_constraintRight_toRightOf="parent" />
+        app:layout_constraintRight_toRightOf="parent"
+        android:onClick="intentChooseTypeOfGame"
+        />
 
     <Button
-        android:text="Button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/button4"
+        android:text="@string/player1"
+        android:layout_width="100dp"
+        android:layout_height="50dp"
+        android:id="@+id/button11"
         android:layout_marginTop="32dp"
-        app:layout_constraintTop_toBottomOf="@+id/button3"
-        app:layout_constraintLeft_toLeftOf="@+id/button3" />
+        app:layout_constraintTop_toBottomOf="@+id/button10"
+        app:layout_constraintLeft_toLeftOf="@+id/button10"
+        android:onClick="chooseDifficultyFragment"/>
 
     <Button
-        android:text="Button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/button10"
+        android:text="@string/highScores"
+        android:layout_width="100dp"
+        android:layout_height="50dp"
+        android:id="@+id/button12"
         android:layout_marginTop="32dp"
-        app:layout_constraintTop_toBottomOf="@+id/button4"
-        app:layout_constraintLeft_toLeftOf="@+id/button4" />
+        app:layout_constraintTop_toBottomOf="@+id/button11"
+        app:layout_constraintLeft_toLeftOf="@+id/button11"
+        android:onClick="intentHighScores"/>
 </android.support.constraint.ConstraintLayout>
